backend/requirements.txt
Django>=3.2,<4.0
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pillow
qrcode[pil]
